Forms: Reduced duplication and boilerplate code on authN forms. 	+ Created utility component HCaptcha.
